# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Zault API
  description: |
    Zault investment dashboard backend API.

    **API versioning:** All endpoints under `/api/**` except `/api/auth/login`, `/api/auth/register`, and `/api/health*` require the `X-API-Version: 1` request header. Missing or unsupported values return `400` with a structured error body.
  version: 0.0.1
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/auth/register:
    post:
      tags:
      - auth-controller
      summary: Register a new user
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "201":
          description: User registered
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          description: Validation error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/logout:
    post:
      tags:
      - auth-controller
      summary: Logout current user
      operationId: logout
      parameters:
      - name: X-API-Version
        in: header
        description: API version — must be `1`
        required: true
        schema:
          type: string
          default: "1"
          enum:
          - "1"
      responses:
        "200":
          description: Logged out
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LogoutResponse"
  /api/auth/login:
    post:
      tags:
      - auth-controller
      summary: Authenticate user
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: Authenticated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Validation error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: Account locked
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/auth/me:
    get:
      tags:
      - auth-controller
      summary: Get current authenticated user
      operationId: me
      parameters:
      - name: X-API-Version
        in: header
        description: API version — must be `1`
        required: true
        schema:
          type: string
          default: "1"
          enum:
          - "1"
      responses:
        "200":
          description: Current user
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Not authenticated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        password:
          type: string
    RegisterResponse:
      type: object
      properties:
        username:
          type: string
        emailVerified:
          type: boolean
        userId:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
    LogoutResponse:
      type: object
      properties:
        status:
          type: string
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        username:
          type: string
        emailVerified:
          type: boolean
        scopes:
          type: array
          items:
            type: string
          uniqueItems: true
...
    MeResponse:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        emailVerified:
          type: boolean
        userId:
          type: string
        scopes:
          type: array
          items:
            type: string
          uniqueItems: true
        userDbPath:
          type: string
        userDbCreatedAt:
          type: string